Add error code class to the mtoken model module Fixed couple missing copyrights Clean up exception handling code
