+++ powerauth-webflow-authentication-mtoken/src/main/java/io/getlime/security/powerauth/lib/webflow/authentication/mtoken/errorhandling/exception/OfflineModeInvalidActivationException.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2017 Lime - HighTech Solutions s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.getlime.security.powerauth.lib.webflow.authentication.mtoken.errorhandling.exception;
+
+import io.getlime.security.powerauth.lib.webflow.authentication.exception.AuthStepException;
+
+/**
+ * Exception for state when activation is missing.
+ *
+ * @author Roman Strobl, roman.strobl@lime-company.eu
+ */
+public class OfflineModeInvalidActivationException extends AuthStepException {
+
+    public OfflineModeInvalidActivationException(String message) {
+        super(message, "offlineMode.invalidActivation");
+    }
+
+}
